LayoutTests/imported/w3c/web-platform-tests/html/semantics/interactive-elements/the-dialog-element/dialog-focus-shadow-expected.txt


FAIL show: When autofocus is not present, the first focusable shadow-including descendant must be focused assert_equals: expected Element node <div></div> but got Element node <body>

<dialog data-description="When autofocus is not p...
FAIL showModal: When autofocus is not present, the first focusable shadow-including descendant must be focused assert_equals: expected Element node <div></div> but got Element node <body>

<dialog data-description="When autofocus is not p...
FAIL show: autofocus outside a shadow tree must take precedence over earlier in-shadow-tree focusable elements assert_equals: expected Element node <div></div> but got Element node <button tabindex="-1" autofocus="">Focusable</button>
FAIL showModal: autofocus outside a shadow tree must take precedence over earlier in-shadow-tree focusable elements assert_equals: expected Element node <div></div> but got Element node <button tabindex="-1" autofocus="">Focusable</button>
FAIL show: autofocus inside a shadow tree must be ignored: no focusable elements outside the shadow tree assert_equals: expected Element node <div></div> but got Element node <body>

<dialog data-description="When autofocus is not p...
FAIL showModal: autofocus inside a shadow tree must be ignored: no focusable elements outside the shadow tree assert_equals: expected Element node <div></div> but got Element node <body>

<dialog data-description="When autofocus is not p...
PASS show: autofocus inside a shadow tree must be ignored: focusable element before the shadow tree
PASS showModal: autofocus inside a shadow tree must be ignored: focusable element before the shadow tree
PASS show: autofocus inside a shadow tree must be ignored: focusable element after the shadow tree
PASS showModal: autofocus inside a shadow tree must be ignored: focusable element after the shadow tree
